fix(ui): restore latestVersion from api/version, fix test shadowing
